To extract logic popup to another component
