+++ src/custom/state/application/atoms.ts
@@ -5,14 +5,14 @@ export { useUpdateAtom, useAtomValue, selectAtom }
 
 type FollowPendingTxPopup = {
   showPopup: boolean
-  hideFollowTxPopup: boolean
+  hidePopupPermanently: boolean
 }
 
 /**
  * Base atom that store the popup state that indicate how to follow a pending tx
  */
 export const followPendingTxPopupAtom = atomWithStorage<FollowPendingTxPopup>('followPendingTxPopup', {
-  hideFollowTxPopup: false,
+  hidePopupPermanently: false,
   showPopup: false,
 })
 
@@ -20,11 +20,11 @@ export const handleFollowPendingTxPopupAtom = atom(null, (_get, set, showPopup: 
   set(followPendingTxPopupAtom, (prev) => ({ ...prev, showPopup }))
 })
 
-export const handleHidePopupPermanentlyAtom = atom(null, (_get, set, hidePopup: boolean) => {
-  set(followPendingTxPopupAtom, (prev) => ({ ...prev, hideFollowTxPopup: hidePopup }))
+export const handleHidePopupPermanentlyAtom = atom(null, (_get, set, hidePopupPermanently: boolean) => {
+  set(followPendingTxPopupAtom, (prev) => ({ ...prev, hidePopupPermanently }))
 })
 
 export const showFollowTxPopupAtom = selectAtom(
   followPendingTxPopupAtom,
-  ({ showPopup, hideFollowTxPopup }) => showPopup && !hideFollowTxPopup
+  ({ showPopup, hidePopupPermanently }) => showPopup && !hidePopupPermanently
 )
